implementacao dos usuarios e suas paginas

+++ app/adm_routes/adm_routes.py
@@ -1,21 +1,9 @@
 from app import app
 from flask import render_template, request, redirect, url_for
-from app.service.auth_service import authenticate
 from app.models.agendamento import Agendamento
 from app.models.estudio import Estudio
-from app.service.auth_service import login_required, current_user
+from flask_login import login_required, current_user
 
-
-@app.route("/adm/login", methods=["GET", "POST"])
-def login():
-    if request.method == "POST":
-        username = request.form.get("username")
-        password = request.form.get("password")
-
-        # Aqui você deve implementar a lógica de autenticação
-        if authenticate(username, password):
-            return redirect(url_for("home"))
-    return render_template("login.html")
 
 
 @app.route("/adm/logout")
@@ -23,7 +11,7 @@ def logout():
     from app.service.auth_service import logout
 
     logout()
-    return redirect(url_for("home"))
+    return redirect(url_for("index"))
 
 
 @app.route("/adm/dashboard")
@@ -37,12 +25,16 @@ def adm_dashboard():
 def delete(agendamento_id):
 
     agendamento = Agendamento.query.get(agendamento_id)
+
     if not agendamento:
         return "Agendamento não encontrado", 404
+    
+    if agendamento.user_id != current_user.id:
+        return "Você não tem permissão para excluir este agendamento", 403
 
     agendamento.delete()
 
-    return redirect(request.referrer or url_for("home"))
+    return redirect(request.referrer or url_for("home", username=current_user.username))
 
 
 @app.route("/estudios/add", methods=["GET", "POST"])
@@ -51,8 +43,8 @@ def add_estudio():
     if request.method == "POST":
         nome = request.form.get("nome")
         descricao = request.form.get("descricao")
-        estudio = Estudio(nome=nome, descricao=descricao)
+        estudio = Estudio(nome=nome, descricao=descricao, user_id=current_user.id)
         estudio.save()
-        return redirect(request.referrer or url_for("home"))
+        return redirect(url_for("home", username=current_user.username))
 
     return render_template("add_estudio.html")
